"[docs]:250925_기본 설정 구성중 7"

+++ src/main/java/com/otd/otd_pointShop/application/point/PointController.java
@@ -91,7 +91,7 @@ public class PointController {
 
     @DeleteMapping("/delete")
     public ResponseEntity<?> deletePointItem(
-            @RequestParam("pointId") Long pointId,
+            @RequestParam("pointId") Integer pointId,
             HttpSession session
     ) {
         Long userId = (Long) session.getAttribute("userId");
